+++ src/GoogleAddressInput/GoogleAddressInput.spec.js
@@ -37,8 +37,8 @@ export class GmapsTestClient {
     }
 
     return Promise.resolve([
-      {description: JSON.stringify(request) + ' - 1'},
-      {description: JSON.stringify(request) + ' - 2'}
+      {description: JSON.stringify(request) + ' - 1', id: 0},
+      {description: JSON.stringify(request) + ' - 2', id: 1}
     ]);
   }
 
@@ -113,7 +113,7 @@ describe('GoogleAddressInput', () => {
     describe('is `true`', () => {
       it('should show google footer', () => {
         const component = createMount({Client: GmapsTestClient, poweredByGoogle: true});
-        component.setState({suggestions: ['a', 'b', 'c'].map(s => ({description: s}))});
+        component.setState({suggestions: ['a', 'b', 'c'].map(s => ({description: s, id: s}))});
         expect(component.find('[data-hook="google-footer"]').exists()).toEqual(true);
       });
 
@@ -133,6 +133,7 @@ describe('GoogleAddressInput', () => {
   });
 
   describe('User Interactions', () => {
+    const defaultSuggestions = [JSON.parse('{"description": "my address", "id": "my-id", "place_id": 123}')];
 
     it('should specify autoSelect as default option', () => {
       const component = createMount({Client: GmapsTestClient, countryCode: 'XX'});
@@ -186,7 +187,7 @@ describe('GoogleAddressInput', () => {
       const onSet = sinon.spy();
 
       const component = createShallow({Client: GmapsTestClient, countryCode: 'XX', onSet});
-      component.setState({suggestions: [JSON.parse('{"description": "my address", "place_id": 123}')]});
+      component.setState({suggestions: defaultSuggestions});
       component.find('InputWithOptions').props().onSelect({id: 0, value: 'my address'});
 
       // Defer to make sure all promises run
@@ -210,7 +211,7 @@ describe('GoogleAddressInput', () => {
         onSet,
         handler: GoogleAddressInputHandler.places
       });
-      component.setState({suggestions: [JSON.parse('{"description": "my address", "place_id": 123}')]});
+      component.setState({suggestions: defaultSuggestions});
       component.find('InputWithOptions').props().onSelect({id: 0, value: 'my address'});
 
       // Defer to make sure all promises run
@@ -228,7 +229,7 @@ describe('GoogleAddressInput', () => {
       const onSet = sinon.spy();
 
       const component = createShallow({Client: GmapsTestClient, countryCode: 'XX', onSet});
-      component.setState({suggestions: [JSON.parse('{"description": "my address", "place_id": 123}')]});
+      component.setState({suggestions: defaultSuggestions});
       component.find('InputWithOptions').props().onManuallyInput('my addr');
 
       // Defer to make sure all promises run
@@ -245,7 +246,7 @@ describe('GoogleAddressInput', () => {
     it('If user pressed <enter> with a value that is not on the suggestions list, try to suggest it and return null if unsuccessful', done => {
       const onSet = sinon.spy();
       const component = createShallow({Client: GmapsTestClient, countryCode: 'YY', onSet});
-      component.setState({suggestions: [JSON.parse('{"description": "my address", "place_id": 123}')]});
+      component.setState({suggestions: defaultSuggestions});
       component.find('InputWithOptions').props().onManuallyInput('dontfind');
 
       // Defer to make sure all promises run
@@ -279,7 +280,7 @@ describe('GoogleAddressInput', () => {
     it('clear suggestions on blur', () => {
       jest.useFakeTimers();
       const component = createMount({Client: GmapsTestClient, countryCode: 'XX', clearSuggestionsOnBlur: true});
-      component.setState({suggestions: [JSON.parse('{"description": "my address", "place_id": 123}')]});
+      component.setState({suggestions: defaultSuggestions});
       component.find('input').simulate('blur');
       jest.runAllTimers();
       component.update();
@@ -289,7 +290,7 @@ describe('GoogleAddressInput', () => {
 
     it('don\'t clear suggestions if clearSuggestionsOnBlur === false', done => {
       const component = createShallow({Client: GmapsTestClient, countryCode: 'XX', clearSuggestionsOnBlur: false});
-      component.setState({suggestions: [JSON.parse('{"description": "my address", "place_id": 123}')]});
+      component.setState({suggestions: defaultSuggestions});
       component.find('InputWithOptions').props().onBlur();
       setTimeout(() => {
         expect(component.find('InputWithOptions').props().options.length).toEqual(1);