+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Types of changes:
 1. **Security** in case of vulnerabilities.
 1. **Breaking** for breaking changes
 
+## 3.2.1 - 2018-06-17
+### Fixed
+- `<GoogleAddressInput>` - Fix id management issue [#1835](https://github.com/wix/wix-style-react/pull/1835)
+
 ## 3.2.0 - 2018-05-28
 ### Added
 - `<MultiSelectCheckbox/>` - new component [#1755](https://github.com/wix/wix-style-react/pull/1755)